Refactoring functional tests based on standard usages
Refactoring the functional tests by using the standard usages [1]
and best pratices available in the oslo.cache documentation.

The main goal of these changes is to help us to
test oslo.cache in the way that it's used in the real life.

It reduce the complexity to test new backends by using the
corresponding configuration.

These changes allow to test a new backend by:
- passing the right environment variables with the related backend to test;
- adding the corresponding zuul job that will dedicated to test this backend.
- adding the right server backend to bindep

[1] https://docs.openstack.org/oslo.cache/latest/user/usage.html

Change-Id: Iaf196d2d93225afa54e324fb830761049059926e