POC to refactor oslo.cache's functional tests
The main goal is to design a testing machinery for functional tests
purpose to load the right config related to the backend to test.

Usage:
------
$ PIFPAF_DAEMON=etcd OSLO_BACKEND=etcd3gw tox -e poc
$ PIFPAF_DAEMON=memcached OSLO_BACKEND=memcache_pool tox -e poc

zuul call functional tests in a similar way.

Important:
----------
The both commands will fails because pifpaf is not called and so
backend is not settled, but the main goal here is simply to check that the
right config is loaded.

Related reviews (ordered by commit order):
--------------------------------------------
- https://review.opendev.org/#/c/712479/
- https://review.opendev.org/#/c/718223/
- https://review.opendev.org/#/c/718176/
- https://review.opendev.org/#/c/711220/

Change-Id: Ib6cc5a31ca87e5c1e45fc2aca4e4c079f541851a